move stdin back to App
