+++ jitpack.yml
@@ -1,7 +1,7 @@
 jdk:
-  - openjdk17
+  - openjdk21
 before_install:
-  - sdk install java 17.0.3-tem
-  - sdk use java 17.0.3-tem
+  - sdk install java 21.0.4-open
+  - sdk use java 21.0.4-open
   - sdk install maven
   - mvn -v
